Moved last block inserted actions test from editor to the block-editor
