+++ packages/block-editor/src/store/test/actions.js
@@ -58,6 +58,8 @@ const {
 	updateSettings,
 	selectionChange,
 	validateBlocksToTemplate,
+	addLastBlockInserted,
+	clearLastBlockInserted,
 } = actions;
 
 describe( 'actions', () => {
@@ -1506,6 +1508,26 @@ describe( 'actions', () => {
 			);
 
 			expect( result ).toEqual( false );
+		} );
+	} );
+
+	describe( 'addLastBlockInserted', () => {
+		it( 'should return the ADD_LAST_BLOCK_INSERTED action', () => {
+			const expectedClientId = 1;
+			const result = addLastBlockInserted( expectedClientId );
+			expect( result ).toEqual( {
+				type: 'ADD_LAST_BLOCK_INSERTED',
+				clientId: expectedClientId,
+			} );
+		} );
+	} );
+
+	describe( 'clearLastBlockInserted', () => {
+		it( 'should return the CLEAR_LAST_BLOCK_INSERTED action', () => {
+			const result = clearLastBlockInserted();
+			expect( result ).toEqual( {
+				type: 'CLEAR_LAST_BLOCK_INSERTED',
+			} );
 		} );
 	} );
 } );
