+++ packages/editor/src/store/test/actions.native.js
@@ -1,11 +1,7 @@
 /**
  * Internal dependencies
  */
-import {
-	togglePostTitleSelection,
-	addLastBlockInserted,
-	clearLastBlockInserted,
-} from '../actions';
+import { togglePostTitleSelection } from '../actions';
 
 describe( 'actions native', () => {
 	describe( 'togglePostTitleSelection', () => {
@@ -14,26 +10,6 @@ describe( 'actions native', () => {
 			expect( result ).toEqual( {
 				type: 'TOGGLE_POST_TITLE_SELECTION',
 				isSelected: true,
-			} );
-		} );
-	} );
-
-	describe( 'addLastBlockInserted', () => {
-		it( 'should return the ADD_LAST_BLOCK_INSERTED action', () => {
-			const expectedClientId = 1;
-			const result = addLastBlockInserted( expectedClientId );
-			expect( result ).toEqual( {
-				type: 'ADD_LAST_BLOCK_INSERTED',
-				clientId: expectedClientId,
-			} );
-		} );
-	} );
-
-	describe( 'clearLastBlockInserted', () => {
-		it( 'should return the CLEAR_LAST_BLOCK_INSERTED action', () => {
-			const result = clearLastBlockInserted();
-			expect( result ).toEqual( {
-				type: 'CLEAR_LAST_BLOCK_INSERTED',
 			} );
 		} );
 	} );
